refactor: tablegroup 패키지 이동
