+++ src/main/java/kitchenpos/domain/OrderTable.java
@@ -12,6 +12,7 @@ import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import kitchenpos.common.domain.NumberOfGuests;
 import kitchenpos.order.domain.Order;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 @Entity
 @Table(name = "order_table")
