+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupRequest.java
@@ -1,7 +1,8 @@
-package kitchenpos.dto;
+package kitchenpos.tablegroup.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import kitchenpos.dto.IdOfOrderTableRequest;
 
 public class TableGroupRequest {
